refactor into one dir
